refactor: replace dynamic [tab] route with explicit page files
- Add dedicated route files: loan.tsx, collateral.tsx, liquidation-price.tsx, expiration.tsx
- Remove dynamic [tab].tsx in favor of Next.js file-based routing
- Update index.tsx route enum for liquidation-price
