+++ pages/mint/[address]/manage/collateral.tsx
@@ -3,40 +3,35 @@ import { useTranslation } from "next-i18next";
 import { useRouter } from "next/router";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { SectionTitle } from "@components/SectionTitle";
-import { AdjustPosition, Target } from "@components/PageMint/AdjustPosition";
+import { AdjustCollateral } from "@components/PageMint/AdjustCollateral";
 import AppCard from "@components/AppCard";
 import { TOKEN_SYMBOL } from "@utils";
 import { usePositionManageData } from "../../../../hooks/usePositionManageData";
 
-enum Route {
-	LOAN = "loan",
-	COLLATERAL = "collateral",
-	LIQUIDATION_PRICE = "liquidation-price",
-	EXPIRATION = "expiration",
-}
-
-export default function PositionManage() {
+export default function ManageCollateral() {
 	const { t } = useTranslation();
 	const router = useRouter();
 	const { address: addressQuery } = router.query;
 
-	const { position, collateralBalance, currentDebt, liqPrice, isInCooldown, cooldownRemainingFormatted, cooldownEndsAt, isLoading } =
-		usePositionManageData(addressQuery);
+	const {
+		position,
+		principal,
+		positionPrice,
+		collateralBalance,
+		currentDebt,
+		minimumCollateral,
+		jusdAllowance,
+		jusdBalance,
+		walletBalance,
+		isInCooldown,
+		cooldownRemainingFormatted,
+		cooldownEndsAt,
+		currentPosition,
+		refetch,
+		isLoading,
+	} = usePositionManageData(addressQuery);
 
-	const handleSelectTarget = (target: Target) => {
-		const targetToRoute: Record<Target, Route> = {
-			[Target.LOAN]: Route.LOAN,
-			[Target.COLLATERAL]: Route.COLLATERAL,
-			[Target.LIQ_PRICE]: Route.LIQUIDATION_PRICE,
-			[Target.EXPIRATION]: Route.EXPIRATION,
-		};
-		const route = targetToRoute[target];
-		if (route) {
-			router.push(`/mint/${addressQuery}/manage/${route}`);
-		}
-	};
-
-	if (isLoading || !position) {
+	if (isLoading || !position || !currentPosition) {
 		return (
 			<div className="md:mt-8 flex justify-center">
 				<AppCard className="max-w-lg w-full p-6 flex flex-col gap-y-6">
@@ -52,22 +47,27 @@ export default function PositionManage() {
 		<>
 			<Head>
 				<title>
-					{TOKEN_SYMBOL} - {t("my_positions.manage_position")}
+					{TOKEN_SYMBOL} - {t("mint.collateral")}
 				</title>
 			</Head>
 			<div className="md:mt-8 flex justify-center">
 				<AppCard className="max-w-lg w-full p-6 flex flex-col gap-y-6">
 					<SectionTitle className="!mb-0 text-center !text-xl">{t("mint.adjust_your_borrowing_position")}</SectionTitle>
-
-					<AdjustPosition
+					<AdjustCollateral
 						position={position}
 						collateralBalance={collateralBalance}
 						currentDebt={currentDebt}
-						liqPrice={liqPrice}
-						onSelectTarget={handleSelectTarget}
+						positionPrice={positionPrice}
+						principal={principal}
+						walletBalance={walletBalance}
+						minimumCollateral={minimumCollateral}
+						jusdBalance={jusdBalance}
+						jusdAllowance={jusdAllowance}
+						refetchAllowance={refetch}
 						isInCooldown={isInCooldown}
 						cooldownRemainingFormatted={cooldownRemainingFormatted}
 						cooldownEndsAt={cooldownEndsAt}
+						onSuccess={refetch}
 					/>
 				</AppCard>
 			</div>
