+++ pages/mint/[address]/manage/loan.tsx
@@ -3,40 +3,33 @@ import { useTranslation } from "next-i18next";
 import { useRouter } from "next/router";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { SectionTitle } from "@components/SectionTitle";
-import { AdjustPosition, Target } from "@components/PageMint/AdjustPosition";
+import { AdjustLoan } from "@components/PageMint/AdjustLoan";
 import AppCard from "@components/AppCard";
 import { TOKEN_SYMBOL } from "@utils";
 import { usePositionManageData } from "../../../../hooks/usePositionManageData";
 
-enum Route {
-	LOAN = "loan",
-	COLLATERAL = "collateral",
-	LIQUIDATION_PRICE = "liquidation-price",
-	EXPIRATION = "expiration",
-}
-
-export default function PositionManage() {
+export default function ManageLoan() {
 	const { t } = useTranslation();
 	const router = useRouter();
 	const { address: addressQuery } = router.query;
 
-	const { position, collateralBalance, currentDebt, liqPrice, isInCooldown, cooldownRemainingFormatted, cooldownEndsAt, isLoading } =
-		usePositionManageData(addressQuery);
+	const {
+		position,
+		principal,
+		collateralBalance,
+		currentDebt,
+		liqPrice,
+		walletBalance,
+		jusdAllowance,
+		isInCooldown,
+		cooldownRemainingFormatted,
+		cooldownEndsAt,
+		currentPosition,
+		refetch,
+		isLoading,
+	} = usePositionManageData(addressQuery);
 
-	const handleSelectTarget = (target: Target) => {
-		const targetToRoute: Record<Target, Route> = {
-			[Target.LOAN]: Route.LOAN,
-			[Target.COLLATERAL]: Route.COLLATERAL,
-			[Target.LIQ_PRICE]: Route.LIQUIDATION_PRICE,
-			[Target.EXPIRATION]: Route.EXPIRATION,
-		};
-		const route = targetToRoute[target];
-		if (route) {
-			router.push(`/mint/${addressQuery}/manage/${route}`);
-		}
-	};
-
-	if (isLoading || !position) {
+	if (isLoading || !position || !currentPosition) {
 		return (
 			<div className="md:mt-8 flex justify-center">
 				<AppCard className="max-w-lg w-full p-6 flex flex-col gap-y-6">
@@ -52,19 +45,24 @@ export default function PositionManage() {
 		<>
 			<Head>
 				<title>
-					{TOKEN_SYMBOL} - {t("my_positions.manage_position")}
+					{TOKEN_SYMBOL} - {t("mint.loan_amount")}
 				</title>
 			</Head>
 			<div className="md:mt-8 flex justify-center">
 				<AppCard className="max-w-lg w-full p-6 flex flex-col gap-y-6">
 					<SectionTitle className="!mb-0 text-center !text-xl">{t("mint.adjust_your_borrowing_position")}</SectionTitle>
-
-					<AdjustPosition
+					<AdjustLoan
 						position={position}
 						collateralBalance={collateralBalance}
 						currentDebt={currentDebt}
 						liqPrice={liqPrice}
-						onSelectTarget={handleSelectTarget}
+						principal={principal}
+						currentPosition={currentPosition}
+						walletBalance={walletBalance}
+						jusdAllowance={jusdAllowance}
+						refetchAllowance={refetch}
+						onSuccess={refetch}
+						onFullRepaySuccess={() => router.push("/dashboard")}
 						isInCooldown={isInCooldown}
 						cooldownRemainingFormatted={cooldownRemainingFormatted}
 						cooldownEndsAt={cooldownEndsAt}
